+++ CHANGELOG.rst
@@ -5,6 +5,11 @@
 Changelog
 =========
 
+* 1.1.0 [2016-05-15]: Added incfile, ste and term_echo functions. These produce
+  output marked up in reStructuredText of source files (incfile) or terminal
+  commands (ste, term_echo). All can be used to include relevant information in
+  docstrings to enhance documentation
+
 * 1.0.5 [2016-05-13]: Minor documentation update
 
 * 1.0.4 [2016-05-02]: Minor documentation and testing enhancements